Improvements fromString
fromString now tries to guess whether an expression is a function or a
variable instead of always taking something with the name of a function
to be a function

fromString now understands - signs better, e.g.  frost('-y') doesn't
return errors. Introduced the NegNode expression for this. May still
need some testing, also in conjuction with simplifier